refactor: Rename library to library_sc2
